Added app store metadata to the repository (Gradle-play-publisher)
* According to the F-Droid docs the file structure for this kind of
metadata can be used to add metadata as long as the files are included
in a tagged release.
https://f-droid.org/en/docs/All_About_Descriptions_Graphics_and_Screenshots/

* While at the moment we (afaik) do not automatically publish to Google
Play we already use Gradle. So the "Gradle Play Publisher" seems to be a
good choice. (F-Droid just uses the file structure. So there is no
necessity to for us to actually use the tool.)
https://github.com/Triple-T/Gradle-play-publisher#managing-play-store-metadata

* This commit includes:
  * contact-email
  * contact-website
  * title for de-DE and en-GB
  * short-description for de-DE and en-GB
  * full-description for de-DE and en-GB
  * a screenshot for de-DE and en-GB