Cleanup unnecessary variables and constant passing
A lot of constants were passed around in a Map. Whilst other values
were packed into the map that are ignored anyway. So this commit moves
nearly all constants to members and bypasses the unnecessary packing
into a Map reducing the code complexity and hopefully make the code more
readable.